feat: replace bouncing logo loader with fullscreen shimmering grey
