+++ .github/CODEOWNERS
@@ -795,7 +795,7 @@ playwright.storybook.config.ts @grafana/grafana-frontend-platform
 /public/app/core/components/AppChrome/ @grafana/grafana-search-navigate-organise
 /public/app/core/components/Branding/ @grafana/grafana-frontend-platform
 /public/app/core/components/Breadcrumbs/ @grafana/grafana-search-navigate-organise
-/public/app/core/components/BouncingLoader/ @grafana/grafana-frontend-platform
+/public/app/core/components/FullPageSkeleton/ @grafana/grafana-frontend-platform
 /public/app/core/components/CardButton.tsx @grafana/dashboards-squad
 /public/app/core/components/CloseButton/ @grafana/grafana-frontend-platform
 /public/app/core/components/ColorScale/ @grafana/dataviz-squad
