+++ public/app/core/navigation/GrafanaRouteLoading.tsx
@@ -3,14 +3,14 @@ import { css } from '@emotion/css';
 import { GrafanaTheme2 } from '@grafana/data';
 import { useStyles2 } from '@grafana/ui';
 
-import { BouncingLoader } from '../components/BouncingLoader/BouncingLoader';
+import { FullPageSkeleton } from '../components/FullPageSkeleton/FullPageSkeleton';
 
 export function GrafanaRouteLoading() {
   const styles = useStyles2(getStyles);
 
   return (
     <div className={styles.loadingPage}>
-      <BouncingLoader />
+      <FullPageSkeleton />
     </div>
   );
 }
@@ -19,9 +19,8 @@ const getStyles = (theme: GrafanaTheme2) => ({
   loadingPage: css({
     backgroundColor: theme.colors.background.primary,
     flex: 1,
-    flexDrection: 'column',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
+    width: '100%',
+    height: '100%',
+    overflow: 'hidden',
   }),
 });
