+++ public/locales/pl-PL/grafana.json
@@ -3544,7 +3544,7 @@
       "tip": "Najedź kursorem na dowolny element w menu nawigacyjnym i kliknij ikonę zakładki, aby dodać ją tutaj."
     }
   },
-  "bouncing-loader": {
+  "full-page-skeleton": {
     "label": "Ładowanie"
   },
   "browse-dashboards": {
