+++ public/locales/ru-RU/grafana.json
@@ -3544,7 +3544,7 @@
       "tip": "Наведите курсор на любой элемент в меню навигации и нажмите значок закладки, чтобы добавить его."
     }
   },
-  "bouncing-loader": {
+  "full-page-skeleton": {
     "label": "Загрузка"
   },
   "browse-dashboards": {
