+++ public/locales/tr-TR/grafana.json
@@ -3520,7 +3520,7 @@
       "tip": "Gezinme menüsündeki herhangi bir ögenin üzerine gelin ve yer imi simgesine tıklayarak buraya ekleyin."
     }
   },
-  "bouncing-loader": {
+  "full-page-skeleton": {
     "label": "Yükleniyor"
   },
   "browse-dashboards": {
