+++ public/views/index.html
@@ -43,15 +43,18 @@
   <body class="theme-[[ .ThemeType ]] [[.AppNameBodyClass]]">
     <style>
       .preloader {
-        height: 100%;
-        flex-direction: column;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin: 0 32px;
+        height: 100vh;
+        width: 100vw;
+        position: fixed;
+        inset: 0;
+        margin: 0;
+        display: block;
+        overflow: hidden;
       }
 
       .preloader__enter {
+        width: 100%;
+        height: 100%;
         opacity: 0;
         animation-name: preloader-fade-in;
         animation-iteration-count: 1;
@@ -60,23 +63,29 @@
         animation-fill-mode: forwards;
       }
 
-      .preloader__bounce {
-        text-align: center;
-        animation-name: preloader-bounce;
-        animation-duration: 0.9s;
+      .preloader__pulse {
+        --base-color: #ebebeb;
+        --highlight-color: #f5f5f5;
+        display: block;
+        width: 100%;
+        height: 100%;
+        background: var(--base-color);
+        background-image: linear-gradient(
+          90deg,
+          var(--base-color) 0%,
+          var(--highlight-color) 50%,
+          var(--base-color) 100%
+        );
+        background-size: 200% 100%;
+        animation-name: preloader-shimmer;
+        animation-duration: 1.5s;
         animation-iteration-count: infinite;
+        animation-timing-function: ease-in-out;
       }
 
-      .preloader__logo {
-        display: inline-block;
-        animation-name: preloader-squash;
-        animation-duration: 0.9s;
-        animation-iteration-count: infinite;
-        width: 60px;
-        height: 60px;
-        background-repeat: no-repeat;
-        background-size: contain;
-        background-image: url('[[.LoadingLogo]]');
+      .theme-dark .preloader__pulse {
+        --base-color: #30343c;
+        --highlight-color: #424752;
       }
 
       .preloader__text {
@@ -110,38 +119,18 @@
         }
       }
 
-      @keyframes preloader-bounce {
-        from,
-        to {
-          transform: translateY(0px);
-          animation-timing-function: cubic-bezier(0.3, 0, 0.1, 1);
+      @keyframes preloader-shimmer {
+        0% {
+          background-position: -200% 0;
         }
-        50% {
-          transform: translateY(-50px);
-          animation-timing-function: cubic-bezier(0.9, 0, 0.7, 1);
+        100% {
+          background-position: 200% 0;
         }
       }
 
-      @keyframes preloader-squash {
-        0% {
-          transform: scaleX(1.3) scaleY(0.8);
-          animation-timing-function: cubic-bezier(0.3, 0, 0.1, 1);
-        }
-        15% {
-          transform: scaleX(0.75) scaleY(1.25);
-          animation-timing-function: cubic-bezier(0, 0, 0.7, 0.75);
-        }
-        55% {
-          transform: scaleX(1.05) scaleY(0.95);
-          animation-timing-function: cubic-bezier(0.9, 0, 1, 1);
-        }
-        95% {
-          transform: scaleX(0.75) scaleY(1.25);
-          animation-timing-function: cubic-bezier(0, 0, 0, 1);
-        }
-        100% {
-          transform: scaleX(1.3) scaleY(0.8);
-          animation-timing-function: cubic-bezier(0, 0, 0.7, 1);
+      @media (prefers-reduced-motion: reduce) {
+        .preloader__pulse {
+          animation: none;
         }
       }
 
@@ -151,13 +140,12 @@
       }
 
       /* stop logo animation */
-      .preloader--done .preloader__bounce,
-      .preloader--done .preloader__logo {
+      .preloader--done .preloader__pulse {
         animation-name: none;
         display: none;
       }
 
-      .preloader--done .preloader__logo,
+      .preloader--done .preloader__pulse,
       .preloader--done .preloader__text {
         display: none;
         color: #ff5705 !important;
@@ -204,9 +192,7 @@
 
     <div class="preloader">
       <div class="preloader__enter">
-        <div class="preloader__bounce">
-          <div class="preloader__logo" aria-live="polite" role="status" aria-label="Loading Grafana"></div>
-        </div>
+        <div class="preloader__pulse" aria-live="polite" role="status" aria-label="Loading Grafana"></div>
       </div>
       <div class="preloader__text preloader__text--fail">
         <main>
